Изменила структуру главной страницы
